% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_calculation.R
\name{biomass_calculation}
\alias{biomass_calculation}
\title{Biomass calculation}
\usage{
biomass_calculation(para, land_required)
}
\arguments{
\item{para}{A JSON file}

\item{land_required}{A dataframe computed using the `land_requirement` function}
}
\value{
list
}
\description{
It computes biomass.
}
\examples{
\dontrun{
data(mufindi)
feed_basket_quality <- feed_quality(para)
energy_required <- energy_requirement(para,feed_basket_quality)
land_required <- land_requirement(feed_basket_quality, energy_required, para)
soil_erosion <- soil_health(para, land_required)
nitrogen_balance <- n_balance(para, land_required, soil_erosion)
livestock_productivity <- land_productivity(para)
economics <- economics_payback(para, energy_required)
biomass_calculation(para, land_required)
}

}
