% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differences.R
\name{calculate_differences}
\alias{calculate_differences}
\title{Calculate Differences}
\usage{
calculate_differences(iDir)
}
\arguments{
\item{iDir}{A source directory where output files are stored}
}
\value{
dataframe
}
\description{
It computes difference in environmental impact between scenarios
}
\examples{
\dontrun{
calculate_differences(iDir)
}

}
