% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_scenario.R
\name{compare_scenario}
\alias{compare_scenario}
\title{Compare Scenarios}
\usage{
compare_scenario(baseRun, outFile, oDir2)
}
\arguments{
\item{baseRun}{name given to the base run}

\item{outFile}{Path to the comparison json computed using the `differences` function}

\item{oDir2}{directory where the plots are to be saved}
}
\description{
It plots the change in scenarios to base run
}
\examples{
\dontrun{
data(mufindi)
data(ghg_para)
feed_basket_quality <- feed_quality(para)
energy_required <- energy_requirement(para,feed_basket_quality)
land_required <- land_requirement(feed_basket_quality, energy_required, para)
soil_erosion <- soil_health(para, land_required)
water_required <- water_requirement(para,land_required)
nitrogen_balance <- n_balance(para, land_required, soil_erosion)
livestock_productivity <- land_productivity(para)
biomass <- biomass_calculation(para, land_required)
soil_carbon <- soil_organic_carbon(para, land_required, biomass)
ghg_emissions <- ghg_emission(para,energy_required,ghg_para,land_required,nitrogen_balance)
combineOutputs(para,feed_basket_quality,energy_required,land_required,soil_erosion,water_required,
nitrogen_balance,livestock_productivity,economics,biomass,soil_carbon,ghg_emissions,filePath)
calculate_differences(outFile,...)
clean_plotting(outFile,oDir)
compare_scenario(baseRun, outFile, oDir2)
}

}
