% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economics_payback.R
\name{economics_payback}
\alias{economics_payback}
\title{Economics payback}
\usage{
economics_payback(para, energy_required, livestock_productivity)
}
\arguments{
\item{para}{A JSON file}

\item{energy_required}{A list computed using the `energy_required` function}

\item{livestock_productivity}{A dataframe containing productivity values.}
}
\value{
dataframe
}
\description{
It computes economic payback from livestock products.
}
\examples{
\dontrun{
data(mufindi)
feed_basket_quality <- feed_quality(para)
energy_required <- energy_requirement(para,feed_basket_quality)
land_required <- land_requirement(feed_basket_quality, energy_required, para)
soil_erosion <- soil_health(para, land_required)
nitrogen_balance <- nitrogen_balance(para, land_required, soil_erosion)
livestock_productivity <- land_productivity(para)
economics_payback(para, energy_required)
}

}
