% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_requirement.R
\name{energy_requirement}
\alias{energy_requirement}
\title{Energy requirement}
\usage{
energy_requirement(para, feed_basket_quality)
}
\arguments{
\item{para}{A json file}

\item{feed_basket_quality}{A dataframe computed using the `feed_quality` function}
}
\value{
dataframe
}
\description{
It calculates energy requirement. Compute this after
the `feed_quality` function
}
\examples{
\dontrun{
data(mufindi)
feed_basket_quality <- feed_quality(para)
energy_requirement(para,feed_basket_quality)
}

}
