% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed_quality.R
\name{feed_quality}
\alias{feed_quality}
\title{Feed quality}
\usage{
feed_quality(para)
}
\arguments{
\item{para}{A list}
}
\value{
dataframe
}
\description{
It computes feed quality
}
\examples{
\dontrun{
data(mufindi)
feed_quality(mufindi)
}

}
