% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_emission.R
\name{ghg_emission}
\alias{ghg_emission}
\title{Greenhouse gas emissions}
\usage{
ghg_emission(
  para,
  energy_required,
  ghg_ipcc_data,
  land_required,
  nitrogen_balance
)
}
\arguments{
\item{para}{A JSON file containing user inputs}

\item{energy_required}{A list computed using the `energy_required` function}

\item{ghg_ipcc_data}{A JSON file containing IpCC coeficients}

\item{land_required}{A list computed using the `land_requirement` function}

\item{nitrogen_balance}{A dataframe computed using the `n_balance` function}
}
\value{
list
}
\description{
It computes greenhouse gas emissions.
}
\examples{
\dontrun{
data(mufindi)
data(ghg_para)
feed_basket_quality <- feed_quality(para)
energy_required <- energy_requirement(para,feed_basket_quality)
land_required <- land_requirement(feed_basket_quality, energy_required, para)
soil_erosion <- soil_health(para, land_required)
nitrogen_balance <- n_balance(para, land_required, soil_erosion)
livestock_productivity <- land_productivity(para)
economics <- economics_payback(para, energy_required)
biomass <- biomass_calculation(para, land_required)
ghg_emission(para,energy_required,ghg_ipcc_data,land_required,nitrogen_balance)
}

}
