% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ghg_para}
\alias{ghg_para}
\title{Greenhouse gas parameters for use in cleaned model.}
\format{
A json file contain fourteen lists:
\describe{
  \item{livestock_category_name}{Livestock category name, a character}
  \item{Urinary_energy_frac}{Urinary energy fraction, a number}
  ...
}
}
\usage{
ghg_para
}
\description{
A dataset containing greenhosue gas parameters for use
the cleaned model.
}
\keyword{datasets}
