% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_productivity.R
\name{land_productivity}
\alias{land_productivity}
\title{Land productivity}
\usage{
land_productivity(para)
}
\arguments{
\item{para}{A json file}
}
\value{
dataframe
}
\description{
It computes land productivity. Products include milk and meat.
}
\examples{
\dontrun{
data(mufindi)
feed_basket_quality <- feed_quality(para)
energy_required <- energy_requirement(para,feed_basket_quality)
land_required <- land_requirement(feed_basket_quality, energy_required, para)
soil_erosion <- soil_health(para, land_required)
nitrogen_balance <- nitrogen_balance(para, land_required, soil_erosion)
land_productivity(para)
}

}
