% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_requirement.R
\name{land_requirement}
\alias{land_requirement}
\title{Land requirement}
\usage{
land_requirement(feed_basket_quality, energy_required, para)
}
\arguments{
\item{feed_basket_quality}{A dataframe computed using the `feed_quality` function}

\item{energy_required}{A list computed using the `energy_required` function}

\item{para}{A JSON file}
}
\value{
dataframe
}
\description{
It compute land requirements
}
\examples{
\dontrun{
data(mufindi)
feed_basket_quality <- feed_quality(para)
energy_required <- energy_requirement(para,feed_basket_quality)
land_requirement(feed_basket_quality, energy_required, para)
}

}
