% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mufindi}
\alias{mufindi}
\title{Sample input data for cleaned model.}
\format{
A json file containing several lists:
\describe{
  \item{farm_code}{Farm code, a character}
  \item{farm_name}{Farm name, a character}
  ...
}
}
\usage{
data(mufindi)
}
\description{
A dataset containing livestock and feed variables for
the cleaned model.
}
\examples{
data(mufindi)
farm_code<-mufindi$farm_code
farm_name <- mufindi$farm_name

}
\keyword{datasets}
