% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrogen_balance.R
\name{n_balance}
\alias{n_balance}
\title{N balance}
\usage{
n_balance(para, land_required, soil_erosion)
}
\arguments{
\item{para}{A json file}

\item{land_required}{A dataframe computed using the `land_requirement` function}

\item{soil_erosion}{A dataframe computed using the `soil_health` function}
}
\value{
dataframe
}
\description{
It computes nitrogen balance
}
\examples{
\dontrun{
data(mufindi)
feed_basket_quality <- feed_quality(para)
energy_required <- energy_requirement(para,feed_basket_quality)
land_required <- land_requirement(feed_basket_quality, energy_required, para)
soil_erosion <- soil_health(para, land_required)
n_balance(para, land_required, soil_erosion)
}

}
