% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output}
\alias{output}
\title{JSON output}
\usage{
output(
  feed_basket_quality,
  energy_required,
  land_required,
  soil_erosion,
  water_requirements,
  nitrogen_balance,
  livestock_productivity,
  economics,
  biomass,
  ghg_emissions
)
}
\arguments{
\item{feed_basket_quality}{A dataframe computed using the `feed_quality` function}

\item{energy_required}{A list computed using the `energy_required` function}

\item{land_required}{A dataframe computed using the `land_requirement` function}

\item{soil_erosion}{A dataframe computed using the `soil_health` function}

\item{water_requirements}{A dataframe computed using the `water_requirement` function}

\item{nitrogen_balance}{A dataframe computed using the `n_balance` function}

\item{livestock_productivity}{A dataframe computed using the `land_productivity` function}

\item{economics}{A dataframe computed using the `economics_payback` function}

\item{biomass}{A dataframe computed using the `biomass_calculation` function}

\item{ghg_emissions}{A dataframe computed using the `n_balance` ghg_emission}
}
\value{
saved json file
}
\description{
It generates a json file of all CLEANED_XtRa computation.
}
\examples{
\dontrun{
data(mufindi)
data(ghg_para)
feed_basket_quality <- feed_quality(para)
energy_required <- energy_requirement(para,feed_basket_quality)
land_required <- land_requirement(feed_basket_quality, energy_required, para)
soil_erosion <- soil_health(para, land_required)
water_requirements <- water_requirement(para,land_required)
nitrogen_balance <- n_balance(para, land_required, soil_erosion)
livestock_productivity <- land_productivity(para)
economics <- economics_payback(para, energy_required)
biomass <- biomass_calculations(para, land_required)
ghg_emission <- ghg_emission(para,energy_required,ghg_ipcc_data,land_required,nitrogen_balance)
output(feed_basket_quality,energy_required,land_required,soil_erosion,water_requirements,
nitrogen_balance,livestock_productivity,economics,biomass,ghg_emissions)
}

}
