% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_health.R
\name{soil_health}
\alias{soil_health}
\title{Soil health}
\usage{
soil_health(para, land_required)
}
\arguments{
\item{para}{A json file}

\item{land_required}{A dataframe computed using the `land_requirement` function}
}
\value{
dataframe
}
\description{
It computes soil health
}
\examples{
\dontrun{
data(mufindi)
feed_basket_quality <- feed_quality(para)
energy_required <- energy_requirement(para,feed_basket_quality)
land_required <- land_requirement(feed_basket_quality, energy_required, para)
soil_health(para, land_required)
}

}
