% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc.R, R/soc_depreciated.R
\name{soil_organic_carbon}
\alias{soil_organic_carbon}
\title{Soil Organic Carbon}
\usage{
soil_organic_carbon(para, stock_change_para, land_required, biomass)

soil_organic_carbon(para, stock_change_para, land_required, biomass)
}
\arguments{
\item{para}{A JSON file}

\item{stock_change_para}{A JSON file with stock change parameters}

\item{land_required}{A dataframe computed using the `land_requirement` function}

\item{biomass}{A dataframe computed using the `biomass_calculation` function}
}
\value{
dataframe

dataframe
}
\description{
It computes soil organic carbon status

It computes soil organic carbon status
}
\examples{
\dontrun{
data(mufindi)
feed_basket_quality <- feed_quality(para)
energy_required <- energy_requirement(para,feed_basket_quality)
land_required <- land_requirement(feed_basket_quality, energy_required, para)
soil_erosion <- soil_health(para, land_required)
nitrogen_balance <- n_balance(para, land_required, soil_erosion)
livestock_productivity <- land_productivity(para)
economics <- economics_payback(para, energy_required)
biomass <- biomass_calculation(para, land_required)
soil_organic_carbon(para, stock_change_para, land_required, biomass)
}

\dontrun{
data(mufindi)
feed_basket_quality <- feed_quality(para)
energy_required <- energy_requirement(para,feed_basket_quality)
land_required <- land_requirement(feed_basket_quality, energy_required, para)
soil_erosion <- soil_health(para, land_required)
nitrogen_balance <- n_balance(para, land_required, soil_erosion)
livestock_productivity <- land_productivity(para)
economics <- economics_payback(para, energy_required)
biomass <- biomass_calculation(para, land_required)
soil_organic_carbon(para, stock_change_para, land_required, biomass)
}

}
