% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock_change_para}
\alias{stock_change_para}
\title{Relative stock change parameters for use in cleaned model.}
\format{
A nested list:
\describe{
  \item{cropland}{Cropland factors, a dataframe}
  \item{grassland}{Grassland factors, a dataframe}
  ...
}

#' @keywords datasets
}
\usage{
data(stock_change_para)
}
\description{
A dataset containing relative stock change parameters for use
the cleaned model.
}
\examples{
data(stock_change_para)
cropland_factors<-stock_change_para$cropland
grassland_factors <- stock_change_para$grassland

}
\keyword{datasets}
