% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_requirement.R
\name{water_requirement}
\alias{water_requirement}
\title{Water requirement}
\usage{
water_requirement(para, land_required)
}
\arguments{
\item{para}{A JSON file containing user inputs}

\item{land_required}{A list computed using the `land_requirement` function}
}
\value{
list
}
\description{
It computes water requirement
}
\examples{
\dontrun{
data(mufindi)
feed_basket_quality <- feed_quality(para)
energy_required <- energy_requirement(para,feed_basket_quality)
land_required <- land_requirement(feed_basket_quality, energy_required, para)
soil_erosion <- soil_health(para, land_required)
water_requirement(para, land_required)
}

}
