% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_outputs.R
\name{combineOutputs}
\alias{combineOutputs}
\title{JSON output}
\usage{
combineOutputs(
  feed_basket_quality,
  energy_required,
  land_required,
  soil_erosion,
  water_required,
  nitrogen_balance,
  livestock_productivity,
  economics,
  biomass,
  ghg_emission
)
}
\arguments{
\item{feed_basket_quality}{A dataframe computed using the `feed_quality` function}

\item{energy_required}{A list computed using the `energy_required` function}

\item{land_required}{A dataframe computed using the `land_requirement` function}

\item{soil_erosion}{A dataframe computed using the `soil_health` function}

\item{water_required}{A dataframe computed using the `water_required` function}

\item{nitrogen_balance}{A dataframe computed using the `n_balance` function}

\item{livestock_productivity}{A dataframe computed using the `land_productivity` function}

\item{economics}{A dataframe computed using the `economics_payback` function}

\item{biomass}{A dataframe computed using the `biomass_calculation` function}

\item{ghg_emission}{A dataframe computed using the `n_balance` ghg_emission}
}
\value{
saved json file
}
\description{
It generates a json file of all CLEANED_XtRa computation.
}
\examples{
\dontrun{
data(mufindi)
data(ghg_para)
feed_basket_quality <- feed_quality(mufindi)
energy_required <- energy_requirement(mufindi,feed_basket_quality)
land_required <- land_requirement(feed_basket_quality, energy_required, mufindi)
soil_erosion <- soil_health(mufindi, land_required)
water_requirement <- water_requirement(mufindi,land_required)
nitrogen_balance <- n_balance(mufindi, land_required, soil_erosion)
livestock_productivity <- land_productivity(mufindi)
economics <- economics_payback(mufindi, energy_required)
biomass <- biomass_calculation(mufindi, land_required)
ghg_emission <- ghg_emission(mufindi,energy_required,ghg_para,land_required,nitrogen_balance)
combineOutputs(feed_basket_quality,energy_required,land_required,soil_erosion,water_required,
nitrogen_balance,livestock_productivity,economics,biomass,ghg_emission)
}

}
