% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ghg_para}
\alias{ghg_para}
\title{Greenhouse gas parameters for use in cleaned model.}
\format{
A nested list:
\describe{
  \item{fertilizer_table}{Livestock category name, a dataframe}
  \item{table_2.5}{Urinary energy fraction, a dataframe}
  ...
}

#' @keywords datasets
}
\usage{
data(ghg_para)
}
\description{
A dataset containing greenhosue gas parameters for use
the cleaned model.
}
\examples{
data(ghg_para)
fertilizer_table<-ghg_para$fertilizer_table
table_2.5 <- ghg_para$table_2.5

}
\keyword{datasets}
