% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_productivity.R
\name{land_productivity}
\alias{land_productivity}
\title{Land productivity}
\usage{
land_productivity(para)
}
\arguments{
\item{para}{A JSON file which contains an object named \code{livestock} which contains
more than one arrays. Each array can contain several objects. Data elements inside
each array include: \code{livestock_category_code}: Livestock category code, \code{livestock_category_name}:
Livestock category name, \code{herd_composition}: Herd composition (number), \code{annual_growth}:
Average annual growth per animal (kilogram), \code{body_weight}: Average Body weight (kilogram),
\code{birth_interval}: Birth interval (years), \code{carcass_fraction}: Carcass fraction, \code{energy_meatcontent}:
Energy content meat, \code{protein_meatcontent}: Protein content meat (percent), \code{annual_milk}:
Average annual milk (kilogram), \code{fat_content}: Fat content milk (percent), \code{protein_milkcontent}:
Protein content milk (percent).}
}
\value{
dataframe
}
\description{
It computes land productivity. Products include milk and meat.
}
\examples{
\dontrun{
data(mufindi)
feed_basket_quality <- feed_quality(para)
energy_required <- energy_requirement(para,feed_basket_quality)
land_required <- land_requirement(feed_basket_quality, energy_required, para)
soil_erosion <- soil_health(para, land_required)
nitrogen_balance <- nitrogen_balance(para, land_required, soil_erosion)
land_productivity(para)
}

}
